socialNetworkChecker/requirements.txt
beautifulsoup4==4.9.3
certifi==2021.5.30
charset-normalizer==2.0.3
idna==3.2
linkedin-api==2.0.0a5
lxml==4.6.3
python-dotenv==0.18.0
requests==2.26.0
soupsieve==2.2.1
urllib3==1.26.6
